Move factories.py to the root of the app.
